Port of `scatter_add` to ATen with multithreading support.
